Move featurisation to separate folder
